Move LogREST into pod/rest package
